+++ src/main/java/gamz/projects/pharmacyfair/repository/projects/storage/StudentsInvolvedRepository.java
@@ -1,6 +1,5 @@
 package gamz.projects.pharmacyfair.repository.projects.storage;
 
-import gamz.projects.pharmacyfair.model.entity.projects.storage.ProjectStatus;
 import gamz.projects.pharmacyfair.model.entity.projects.storage.StudentsInvolved;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
@@ -8,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface StudentsInvolvedRepository extends CrudRepository<StudentsInvolved,Integer> {
+public interface StudentsInvolvedRepository extends CrudRepository<StudentsInvolved,Long> {
     Optional<StudentsInvolved> findByCode(String code);
 }
